example: Cleanup some setup code and add a Makefile
